// src/core/types/gen/platform/transaction/transaction.ts
export interface GetTransactionByHashResponse {
  /** Metadata about the request */
  metadata: ResponseMetadata | undefined;
  result: Transaction | undefined;
  /** List of errors encountered during the request */
  errors: string[];
}

export interface Transaction {
  /**
   * Unique transaction identifier
   *
   * examples: "0x8f4f...", "tx_123abc", "internal_456def"
   */
  id: string;
  /**
   * Transaction type
   *
   * examples: "transfer", "swap", "contract_call", "mint"
   */
  type: string;
  /**
   * Blockchain chain identifier
   *
   * examples: "1", "137", "56", "42161"
   */
  chainId: string;
  /**
   * Network name
   *
   * examples: "mainnet", "testnet", etc
   */
  network: string;
  /**
   * Block number where transaction was mined
   *
   * examples: 18500000, 45000000, 32000000
   */
  blockNumber: number;
  /**
   * Number of block confirmations
   *
   * examples: 0, 12, 50, 100
   */
  blockConfirmations: number;
  /**
   * Unix timestamp when transaction was mined
   *
   * examples: "2025-05-21T00:00:00Z", "2025-12-31T23:59:59Z"
   */
  minedAt: Date | undefined;
  /**
   * Transaction status
   *
   * examples: "pending", "confirmed", "failed", "dropped"
   */
  status: string;
  /**
   * Transaction hash
   *
   * examples: "0x1f4f7c2d...", "0x8ba1f109...", "0x742d35cc..."
   */
  hash: string;
  /**
   * Transaction direction relative to the queried address
   *
   * examples: "in", "out", "self"
   */
  direction: string;
  /**
   * Source address of the transaction
   *
   * examples: "0x742D35Cc6634C0532925a3b8D404020ae0e4f5f3"
   */
  addressFrom: string;
  /**
   * Destination address of the transaction
   *
   * examples: "0x8ba1f109551bD432803012645Hac136c78baae"
   */
  addressTo: string;
  /**
   * Transaction nonce
   *
   * examples: 0, 42, 1337, 999999
   */
  nonce: number;
  /** List of balance changes caused by this transaction */
  changes: Change[];
  /** Transaction fee information */
  fee: Fee | undefined;
  /** Additional transaction metadata */
  meta: Meta | undefined;
  /**
   * Raw transaction calldata
   *
   * examples: "0x", "0xa9059cbb000000000000000000000000..."
   */
  callData: Uint8Array;
}

export interface Change {
  /** Asset information for this balance change */
  asset: Asset | undefined;
  /**
   * Direction of the balance change relative to the address
   *
   * examples: "in", "out", "mint", "burn"
   */
  direction: string;
  /**
   * Source address of the balance change
   *
   * examples: "0x742D35Cc6634C0532925a3b8D404020ae0e4f5f3", "0x0000000000000000000000000000000000000000"
   */
  addressFrom: string;
  /**
   * Destination address of the balance change
   *
   * examples: "0x8ba1f109551bD432803012645Hac136c78baae", "0x742D35Cc6634C0532925a3b8D404020ae0e4f5f3"
   */
  addressTo: string;
  /**
   * Price per unit of the asset in the specified currency
   *
   * examples: "1.00", "3456.78", "0.000001", "45123.45"
   */
  price: string;
  /**
   * Human-readable quantity of the asset
   *
   * examples: "1.5", "1000.0", "0.001", "999999.999999"
   */
  quantity: string;
  /**
   * Raw value in the asset's smallest unit (wei, satoshi, etc.)
   *
   * examples: "1000000000000000000", "500000000", "123456789012345678901234567890"
   */
  value: string;
}
// ...
export interface Fee {
  /**
   * Total fee value in the smallest unit (wei, gwei, etc.)
   *
   * examples: "1000000000000000000", "500000000", "123456789012345678901234567890"
   */
  value: string;
  /**
   * Fee value in the specified currency (USD, ETH, etc.)
   *
   * examples: "12.45", "0.003", "156.78", "0.000123"
   */
  price: string;
  /** Detailed fee breakdown information */
  details: FeeDetail | undefined;
}
// ...
export interface Meta {
  /**
   * Action performed in the transaction
   *
   * examples: "transfer", "approve", "swap", "mint", "burn"
   */
  action: string;
  /**
   * Name of the contract involved in the transaction
   *
   * examples: "Uniswap V3", "OpenSea", "USDC Token", "Compound"
   */
  contractName: string;
  /**
   * Function selector (first 4 bytes of function signature)
   *
   * examples: "0xa9059cbb", "0x095ea7b3", "0x23b872dd", "0x40c10f19"
   */
  fourbyte: string;
  /**
   * URL to the contract's icon image
   *
   * examples: "https://example.com/icons/contract.png", "https://assets.example.com/uniswap.svg"
   */
  contractIconUrl: string;
  /**
   * Human-readable label for block explorer display
   *
   * examples: "Token Transfer", "Uniswap Swap", "NFT Mint", "Contract Interaction"
   */
  explorerLabel: string;
  /**
   * URL to view transaction details on block explorer
   *
   * examples: "https://etherscan.io/tx/0x123...", "https://polygonscan.com/tx/0x456..."
   */
  explorerUrl: string;
  /** Asset information related to the transaction */
  asset: Asset | undefined;
  /**
   * Quantity of assets involved in the transaction
   *
   * examples: "1.5", "1000.0", "0.001", "999999.999999"
   */
  quantity: string;
  /**
   * Address that was approved to spend tokens
   *
   * examples: "0x742D35Cc6634C0532925a3b8D404020ae0e4f5f3", "0x0000000000000000000000000000000000000000"
   */
  approvalTo: string;
  /**
   * Transaction type classification
   *
   * examples: "transfer", "swap", "approval", "contract_call", "nft_mint"
   */
  type: string;
  /**
   * Transaction subtype for more granular classification
   *
   * examples: "erc20_transfer", "eth_transfer", "nft_sale", "liquidity_add"
   */
  subType: string;
  /**
   * Public-facing subtype for external display
   *
   * examples: "Token Swap", "NFT Purchase", "Liquidity Provision", "Yield Farming"
   */
  publicSubType: string;
}